bk-platform/backend/requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.1
SQLAlchemy==2.0.30
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-multipart==0.0.9
httpx==0.27.0
chromadb==0.5.3
sentence-transformers==2.7.0
